Move Java classes in scala into Java packages
